requirements.txt
peewee
loguru
tmdbsimple
python-dateutil
beautifulsoup4
flask-login